feat: use dma & adc & tim to sample data on 16khz

+++ app/circular_buffer.h
@@ -7,11 +7,11 @@ typedef struct CircularBuffer* ByteCircularBuffer;
 
 ByteCircularBuffer ByteCircularBuffer_Create(uint16_t capacity);
 void ByteCircularBuffer_Destroy(ByteCircularBuffer circularBuffer);
-uint8_t ByteCircularBuffer_GetSize(ByteCircularBuffer circularBuffer);
+uint16_t ByteCircularBuffer_GetSize(ByteCircularBuffer circularBuffer);
 uint8_t ByteCircularBuffer_IsEmpty(ByteCircularBuffer circularBuffer);
-void ByteCircularBuffer_Push(ByteCircularBuffer circularBuffer, uint8_t value);
-void ByteCircularBuffer_PushFromISR(ByteCircularBuffer circularBuffer, uint8_t value);
-uint8_t ByteCircularBuffer_Pop(ByteCircularBuffer circularBuffer);
-uint8_t ByteCircularBuffer_PopFromISR(ByteCircularBuffer circularBuffer);
+void ByteCircularBuffer_Push(ByteCircularBuffer circularBuffer, uint16_t value);
+void ByteCircularBuffer_PushFromISR(ByteCircularBuffer circularBuffer, uint16_t value);
+uint16_t ByteCircularBuffer_Pop(ByteCircularBuffer circularBuffer);
+uint16_t ByteCircularBuffer_PopFromISR(ByteCircularBuffer circularBuffer);
 
 #endif
